wip/frontend/requirements.txt
Flask
PyJWT
requests
cryptography
